Fix compilation errors, tests, and code quality issues
- Fix type mismatch: convert max_breadcrumbs from u32 to usize
- Fix lifetime issue: use Cow::Owned for environment field
- Fix clippy warnings: use or_default(), remove unnecessary cast, allow too_many_arguments
- Fix test: make duplicate grouping test order-independent
- Fix formatting: remove trailing whitespace
- Fix SBOM warnings: use proper author format (name only)
- Add error tracking module with Sentry/GlitchTip support
- Update documentation for error tracking setup

+++ crates/dragonfly-monitor/src/collector.rs
@@ -22,7 +22,7 @@ impl MetricsCollector {
     pub async fn collect(&mut self) -> Result<SystemMetrics> {
         self.system.refresh_all();
 
-        let cpu_usage = self.system.global_cpu_info().cpu_usage() as f32;
+        let cpu_usage = self.system.global_cpu_info().cpu_usage();
         let total_memory = self.system.total_memory();
         let used_memory = self.system.used_memory();
         let total_swap = self.system.total_swap();
